Renamed existing Environment class to GoogleDatastoreEnvironment. Created abstract Environment class.
