+++ src/main/kotlin/com/studyshare/environment/GoogleDatastoreEnvironment.kt
@@ -1,10 +1,10 @@
-package com.studyshare
+package com.studyshare.environment
 
 import com.google.cloud.datastore.Datastore
 import com.google.cloud.datastore.DatastoreOptions
 import com.google.cloud.datastore.Entity
 
-data class Environment(val PRODUCTION: Boolean) {
+data class GoogleDatastoreEnvironment(override val production: Boolean) : Environment(production) {
     private val datastore: Datastore = DatastoreOptions.newBuilder()
         .setProjectId("skillful-fx-446014-k1")
         .setDatabaseId("environment")
@@ -19,28 +19,28 @@ data class Environment(val PRODUCTION: Boolean) {
     }
 
 
-    val MONGODB_CONNECTION_STRING: String =
+    override val mongodbConnectionString: String =
         if (entity != null) {
             entity.getString("MONGODB_CONNECTION_STRING") ?: throw Exception("No connection string found")
         } else {
             System.getenv("MONGODB_CONNECTION_STRING")
         }
 
-    val DATABASE_NAME: String =
+    override val databaseName: String =
         if (entity != null) {
             entity.getString("DATABASE_NAME") ?: throw Exception("No database name found")
         } else {
             System.getenv("DATABASE_NAME")
         }
 
-    val BUCKET_NAME: String =
+    override val bucketName: String =
         if (entity != null) {
             entity.getString("BUCKET_NAME") ?: throw Exception("No bucket name found")
         } else {
             System.getenv("BUCKET_NAME")
         }
 
-    val STATIC_BUCKET_NAME: String =
+    override val staticBucketName: String =
         if (entity != null) {
             entity.getString("STATIC_BUCKET_NAME") ?: throw Exception("No static bucket name found")
         } else {
